Feat. 로그인 유저 정보를 가져오는 ArgumentResolver 구현 - JWT 토큰을 통해 SecurityContext에 저장되어 있는 유저 이메일로 유저 정보를 가져오도록 구현 - 반복 사용을 위해 Argument Resolver로 적용

+++ src/main/java/com/eastshine/looknshop/domain/Product/Product.java
@@ -29,7 +29,7 @@ public class Product extends BaseEntity {
 
     private int price;
 
-    private BigDecimal discountRate;
+    private int discountRate;
 
     private int totalStock;
